requirements.txt
Django==4.2
django-environ
djangorestframework
django-rest-knox